pip_requirements.txt
jugaad-data
pyarrow
pandas
numpy
matplotlib
datetime
python-dateutil
openpyxl
tables
Jinja2
lxml